Time: 42 ms (58.88%), Space: 53.1 MB (34.11%) - LeetHub

+++ 1621-number-of-sets-of-k-non-overlapping-line-segments/1621-number-of-sets-of-k-non-overlapping-line-segments.cpp
@@ -9,7 +9,7 @@ public:
         }
 
         for (int j = 1; j <= k; j++) {
-           long long sum = 0;
+           int sum = 0;
 
             for (int i = 1; i < n; i++) {
